+++ .jest.js
@@ -2,6 +2,11 @@ export default {
     'transform': {},
     'verbose': true,
     'testSequencer': './test/jestTestSequencer.js',
+    'testPathIgnorePatterns': [
+        '/node_modules/',
+        // tests below are intended to be executed manually
+        '.*\\.manual\\.test\\.js$'
+    ],
     'globals': {
         // neptune db that has pre-loaded air routes sample data host and port
         // ex. db-neptune-foo-bar.cluster-abc.us-west-2.neptune.amazonaws.com